all entites for sales are provided

+++ src/main/java/com/avnzor/oms_backend/sales/repository/ShopifyOrderRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ShopifyOrderRepository extends JpaRepository<ShopifyOrder, Integer> {
+public interface ShopifyOrderRepository extends JpaRepository<ShopifyOrder, Long> {
 
     Optional<ShopifyOrder> findFirstByOrderNumber(String orderNumber);
 }
